Display local media controls also when collapsed
Moved LocalMediaControls to the Grid view to make sure it's always
visible even when the grid is collapsed.
